+++ web/app/store/user/reducer.js
@@ -6,12 +6,14 @@ import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayload} from 'progressive-web-sdk/dist/utils/reducer-utils'
 import {setLoggedIn, receiveUserCustomContent} from 'progressive-web-sdk/dist/integration-manager/results'
+import {recieveAccountAddressData} from 'progressive-web-sdk/dist/integration-manager/account/results.js'
 
 const initialState = Immutable.Map()
 
 const userReducer = handleActions({
     [setLoggedIn]: mergePayload,
-    [receiveUserCustomContent]: mergePayload
+    [receiveUserCustomContent]: mergePayload,
+    [recieveAccountAddressData]: mergePayload
 }, initialState)
 
 export default userReducer
